# Speed up boot time in rspec tests