--- a/spec/integration/streams/on_update_message_spec.rb
+++ b/spec/integration/streams/on_update_message_spec.rb
@@ -141,10 +141,8 @@ RSpec.describe PublishingAPI::Consumer do
       allow(GovukError).to receive(:notify)
     end
 
-    schemas = GovukSchemas::Schema.all(schema_type: 'notification')
-    schemas.each_value do |schema|
+    SchemasIterator.each_schema do |schema_name, schema|
       payload = GovukSchemas::RandomExample.new(schema: schema).payload
-      schema_name = payload.dig('schema_name')
       unless %w{travel_advice guide}.include?(schema_name) || schema_name.include?('placeholder')
 
         %w{major minor links republish unpublish}.each do |update_type|