--- a/spec/integration/streams/process_sub_pages_spec.rb
+++ b/spec/integration/streams/process_sub_pages_spec.rb
@@ -330,11 +330,11 @@ RSpec.describe "Process sub-pages for multipart content types" do
     end
 
     SchemasIterator.each_schema do |schema_name, schema|
-      payload = GovukSchemas::RandomExample.new(schema: schema).payload
       if %w{travel_advice guide}.include?(schema_name) && !schema_name.include?('placeholder')
 
         %w{major minor links republish unpublish}.each do |update_type|
           it "handles event for: `#{schema_name}` with no errors for a `#{update_type}` update" do
+            payload = GovukSchemas::RandomExample.new(schema: schema).payload
             message = build(:message, :with_parts, payload: payload, routing_key: "#{schema_name}.#{update_type}")
 
             subject.process(message)