--- a/spec/integration/streams/process_sub_pages_spec.rb
+++ b/spec/integration/streams/process_sub_pages_spec.rb
@@ -334,7 +334,7 @@ RSpec.describe "Process sub-pages for multipart content types" do
 
         %w{major minor links republish unpublish}.each do |update_type|
           it "handles event for: `#{schema_name}` with no errors for a `#{update_type}` update" do
-            payload = GovukSchemas::RandomExample.new(schema: schema).payload
+            payload = SchemasIterator.payload_for(schema_name, schema)
             message = build(:message, :with_parts, payload: payload, routing_key: "#{schema_name}.#{update_type}")
 
             subject.process(message)