--- a/spec/support/schemas_iterator.rb
+++ b/spec/support/schemas_iterator.rb
@@ -1,9 +1,12 @@
 module SchemasIterator
   def self.each_schema
-    schemas = GovukSchemas::Schema.all(schema_type: "notification")
-    schemas.each do |schema_name, value|
+    all_schemas.each do |schema_name, value|
       schema_name = schema_name.split('/')[-3]
       yield schema_name, value
     end
+  end
+
+  def self.all_schemas
+    @all_schemas ||= GovukSchemas::Schema.all(schema_type: "notification")
   end
 end