--- a/spec/support/schemas_iterator.rb
+++ b/spec/support/schemas_iterator.rb
@@ -6,6 +6,14 @@ module SchemasIterator
     end
   end
 
+  def self.payload_for(schema_name, schema)
+    all_payloads[schema_name] ||= GovukSchemas::RandomExample.new(schema: schema).payload
+  end
+
+  def self.all_payloads
+    @all_payloads ||= {}
+  end
+
   def self.all_schemas
     @all_schemas ||= GovukSchemas::Schema.all(schema_type: "notification")
   end