We are generating tests samples dynamically based on the list
of schemas that are available. The process to read all the schemas
is really fast, but generating a sample for each schema is time 
consuming, so we need to speed it up.

I have done the following improvements:

1. Cache the list of schemas
2. Cache each sample for each type of schema
3. Lazy creation of the sample: we only generate a sample when the test runs

Achievments:

1. Reduce load time of  **all suite time in 20 seconds**
2. Reduce load time of **individual executions in 9 seconds**.

### All suite

```
Running via Spring preloader in process 26787
Randomized with seed 52091
....................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................

Finished in 1 minute 7.92 seconds (files took 17.97 seconds to load)
756 examples, 0 failures
```

```
Randomized with seed 45120
....................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................

Finished in 1 minute 6.37 seconds (files took 0.82826 seconds to load)
756 examples, 0 failures

Randomized with seed 45120
```

### For individual tests


```
% rspec './spec/integration/streams/on_update_message_spec.rb[1:8:33]' # PublishingAPI::Consumer all schemas handles event for: `organisations_homepage` with no errors for a `links` update                                         
Running via Spring preloader in process 26996
Run options: include {:ids=>{"./spec/integration/streams/on_update_message_spec.rb"=>["1:8:33"]}}

Randomized with seed 15246
.

Finished in 0.85963 seconds (files took 9.27 seconds to load)
1 example, 0 failures
```

to

```
% rspec './spec/integration/streams/on_update_message_spec.rb[1:8:33]' # PublishingAPI::Consumer all schemas handles event for: `organisations_homepage` with no errors for a `links` update                                         
Running via Spring preloader in process 27088
Run options: include {:ids=>{"./spec/integration/streams/on_update_message_spec.rb"=>["1:8:33"]}}

Randomized with seed 3630
.

Finished in 1.64 seconds (files took 0.50254 seconds to load)
1 example, 0 failures
```